Splitted bindings into conditions and bindings
